# api/openapi.yaml
openapi: 3.1.0
info:
  title: hostd Phase A API
  version: 0.1.0
  description: Authenticated local-only control plane. Workload execution is disabled by default and requires an explicit runtime capability.
servers:
  - url: http://127.0.0.1:7345
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/apps/{appId}/scoped-configuration:
    parameters: [{$ref: '#/components/parameters/AppId'}]
    put:
      operationId: replaceScopedApplicationConfiguration
      security: [{sessionCookie: [], csrfHeader: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReplaceScopedApplicationConfigurationRequest'}}}}
      responses:
        '200': {description: New immutable component-scoped configuration revision with secret values omitted., headers: {Cache-Control: {schema: {type: string, const: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/ApplicationConfiguration'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hostd_session
      description: Opaque random session token. Set HttpOnly, Path=/, SameSite=Strict; Secure is enabled when served over TLS. Only a SHA-256 hash is stored.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with sessionCookie for every non-safe request. Rotated by the authenticated CSRF endpoint; only its hash is stored.
...
  parameters:
    AppId: {name: appId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Problem:
      description: RFC 9457 problem response
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, code, requestId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        requestId: {type: string}
        errors: {type: object, additionalProperties: {type: string}}
...
    ConfigurationEntry: {type: object, required: [key, sensitive], properties: {key: {type: string}, sensitive: {type: boolean}, value: {type: string}, phase: {type: string, enum: [runtime, build, migration]}, targetComponent: {type: string}}}
...
    ApplicationConfiguration: {type: object, required: [revisionNumber, entries], properties: {revisionId: {type: string, format: uuid}, revisionNumber: {type: integer, format: int64, minimum: 0}, formatVersion: {type: integer, enum: [1, 2]}, deploymentPlanRevisionId: {type: string, format: uuid}, deploymentPlanRevisionNumber: {type: integer, format: int64, minimum: 0}, updatedAt: {type: string, format: date-time}, entries: {type: array, items: {$ref: '#/components/schemas/ConfigurationEntry'}}}}
...
    ScopedConfigurationKey: {type: object, additionalProperties: false, required: [key, phase, targetComponent], properties: {key: {type: string, pattern: '^[A-Za-z_][A-Za-z0-9_]*$', maxLength: 128}, phase: {type: string, enum: [runtime, build, migration]}, targetComponent: {type: string, minLength: 1, maxLength: 256}}}
    ScopedConfigurationValueInput: {type: object, additionalProperties: false, required: [key, sensitive, phase, targetComponent, value], properties: {key: {type: string, pattern: '^[A-Za-z_][A-Za-z0-9_]*$', maxLength: 128}, sensitive: {type: boolean}, phase: {type: string, enum: [runtime, build, migration]}, targetComponent: {type: string, minLength: 1, maxLength: 256}, value: {type: string, maxLength: 8192}, preserveStoredSecret: {type: boolean}}}
    ReplaceScopedApplicationConfigurationRequest: {type: object, additionalProperties: false, required: [expectedRevisionNumber, planRevisionId, planRevisionNumber, entries, remove, publicBuildDisclosureAcknowledged], properties: {expectedRevisionNumber: {type: integer, format: int64, minimum: 0}, planRevisionId: {type: string, format: uuid}, planRevisionNumber: {type: integer, format: int64, minimum: 1}, entries: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/ScopedConfigurationValueInput'}}, remove: {type: array, maxItems: 256, items: {$ref: '#/components/schemas/ScopedConfigurationKey'}}, publicBuildDisclosureAcknowledged: {type: boolean}}}
...